cartorder-app/requirements.txt
fastapi==0.61.2
uvicorn==0.12.2
mongoengine==0.23.0
python-multipart==0.0.5
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.2.0
pydantic[email]
